# Swagger API Documentaion.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a great resource for using endpoints of the monitorMyRehab API. We use a general swagger UI for all endpoints. Completed routes doc - users, admins, appointment, healthprofessionalTypes, sessions",
    "version": "1.0.0",
    "title": "MonitorMyRehab API",
    "termsOfService": "https://monitormyrehab.com/terms/",
    "contact": {
      "email": "info@monitormyrehab.com"
    },
    "license": {
      "name": "UNLICENSED",
      "url": "https://monitormyrehab.com/privacy/"
    }
  },
  "host": "api.mmr-meet.click",
  "basePath": "/api/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/sessions/users/{id}/exercise": {
      "put": {
        "tags": [
          "sessions"
        ],
        "summary": "Add an exercise to session",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a health professional's authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the session to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SessionExercise"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/sessions/users/{id}/exercises": {
      "put": {
        "tags": [
          "sessions"
        ],
        "summary": "Add an exercise to session",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a health professional's authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the session to update",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SessionExercise"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
...
    "/exercises/": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get all exercises",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccessArray"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Add exercise",
        "description": "Add exercise",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "title",
            "in": "formData",
            "description": "title of exercise",
            "required": true,
            "type": "string"
          },
          {
            "name": "description",
            "in": "formData",
            "description": "description of exercise",
            "required": true,
            "type": "file"
          },
          {
            "name": "video",
            "in": "formData",
            "description": "video to upload",
            "required": true,
            "type": "file"
          },
          {
            "name": "image",
            "in": "formData",
            "description": "image to upload",
            "required": true,
            "type": "file"
          },
          {
            "name": "audio",
            "in": "formData",
            "description": "audio to upload",
            "required": true,
            "type": "file"
          },
          {
            "name": "thumbImage",
            "in": "formData",
            "description": "thumbImage to upload",
            "required": true,
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/exercises/paginate": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get paginated exercises",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccessArray"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/exercises/count": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get all exercises count",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccessArray"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
    "/exercises/{id}": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get one exercise by id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "a user authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the exercises to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "put": {
        "tags": [
          "exercises"
        ],
        "summary": "Update exercise by id",
        "description": "Update exercises details",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the exercise to get",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated exercise object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exercises"
        ],
        "summary": "Delete exercise with id",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Auth-Token",
            "in": "header",
            "description": "an admin authorization token",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "UID of the exercise to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "$ref": "#/definitions/PostSuccess"
            }
          },
          "403": {
            "description": "Unauthorized Access",
            "schema": {
              "$ref": "#/definitions/AccessError"
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
              "$ref": "#/definitions/ServerError"
            }
          }
        }
      }
    },
...
      "templates/users/{id}/exercise" : {
        "put" : {
          "tags" : [ "templates" ],
          "summary" : "Add an exercise to template",
          "description" : "",
          "consumes" : [ "application/json" ],
          "produces" : [ "application/json" ],
          "parameters" : [ {
            "name" : "Auth-Token",
            "in" : "header",
            "description" : "a health professional's authorization token",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "id",
            "in" : "path",
            "description" : "id of the template to update",
            "required" : true,
            "type" : "string"
          }, {
            "in" : "body",
            "name" : "body",
            "required" : true,
            "schema" : {
              "$ref" : "#/definitions/SessionExercise"
            }
          } ],
          "responses" : {
            "403" : {
              "description" : "Unauthorized Access",
              "schema" : {
                "$ref" : "#/definitions/AccessError"
              }
            },
            "500" : {
              "description" : "Server Error",
              "schema" : {
                "$ref" : "#/definitions/ServerError"
              }
            },
            "200" : {
              "description" : "Success response",
              "schema" : {
                "$ref" : "#/definitions/PostSuccess"
              }
            }
          }
        }
      },
      "templates/users/{id}/exercises" : {
        "put" : {
          "tags" : [ "templates" ],
          "summary" : "Add an exercise to template",
          "description" : "",
          "consumes" : [ "application/json" ],
          "produces" : [ "application/json" ],
          "parameters" : [ {
            "name" : "Auth-Token",
            "in" : "header",
            "description" : "a health professional's authorization token",
            "required" : true,
            "type" : "string"
          }, {
            "name" : "id",
            "in" : "path",
            "description" : "id of the template to update",
            "required" : true,
            "type" : "string"
          }, {
            "in" : "body",
            "name" : "body",
            "required" : true,
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/SessionExercise"
              }
            }
          } ],
          "responses" : {
            "403" : {
              "description" : "Unauthorized Access",
              "schema" : {
                "$ref" : "#/definitions/AccessError"
              }
            },
            "500" : {
              "description" : "Server Error",
              "schema" : {
                "$ref" : "#/definitions/ServerError"
              }
            },
            "200" : {
              "description" : "Success response",
              "schema" : {
                "$ref" : "#/definitions/PostSuccess"
              }
            }
          }
        }
      },
...
  },
  "definitions": {
...
    "AccessError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": "Shows status of request"
        },
        "message": {
          "type": "string",
          "description": "",
          "enum": [
            "Unauthorized Access"
          ]
        },
        "route": {
          "type": "string",
          "description": ""
        }
      }
    },
    "ServerError": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": "Shows status of request"
        },
        "message": {
          "type": "string",
          "description": "",
          "enum": [
            "Server Error"
          ]
        },
        "route": {
          "type": "string",
          "description": ""
        }
      }
    },
...
    "AddUser": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "User email"
        },
        "role": {
          "type": "string",
          "description": "User email",
          "enum": [
            "ind",
            "pro",
            "admin"
          ]
        },
        "dob": {
          "type": "string",
          "description": "User date of birth"
        },
        "status": {
          "type": "string",
          "description": "User status - verified | unverified",
          "enum": [
            "verified",
            "unverified"
          ]
        },
        "photoUrl": {
          "type": "string",
          "description": "User photoUrl"
        },
        "gender": {
          "type": "string",
          "description": "User's gender - M | F | O",
          "enum": [
            "M",
            "F",
            "O"
          ]
        },
        "address": {
          "$ref": "#/definitions/AddUser_address"
        },
        "preferences": {
          "$ref": "#/definitions/AddUser_preferences"
        },
        "phone": {
          "$ref": "#/definitions/AddUser_phone"
        }
      }
    },
...
    "PostSuccess": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": ""
        },
        "body": {
          "type": "object",
          "description": "Body contains expected response. May be array or object",
          "properties": {}
        },
        "message": {
          "type": "string",
          "description": "Success message"
        },
        "route": {
          "type": "string"
        }
      }
    },
    "PostSuccessArray": {
      "type": "object",
      "properties": {
        "status": {
          "type": "boolean",
          "description": ""
        },
        "body": {
          "type": "array",
          "description": "Body contains expected response. May be array or object",
          "items": {
            "$ref": "#/definitions/Object"
          }
        },
        "message": {
          "type": "string",
          "description": "Success message"
        },
        "route": {
          "type": "string"
        }
      }
    },
    "Object": {
      "type": "object",
      "properties": {
        "generic": {
          "type": "string",
          "description": "Normal JSON Object"
        }
      }
    },
...
    "SessionExercise": {
      "type": "object",
      "properties": {
        "exercise": {
          "type": "string"
        },
        "sets": {
          "type": "integer"
        },
        "time": {
          "type": "integer"
        },
        "distance": {
          "type": "integer"
        },
        "note": {
          "type": "string"
        },
        "intensity": {
          "type": "string"
        }
      }
    },
...
    "AddUser_address": {
      "type": "object",
      "properties": {
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      },
      "description": ""
    },
    "AddUser_preferences": {
      "type": "object",
      "properties": {
        "timezone": {
          "type": "string"
        },
        "country": {
          "type": "string"
        }
      },
      "description": ""
    },
    "AddUser_phone": {
      "type": "object",
      "properties": {
        "dailCode": {
          "type": "string"
        },
        "number": {
          "type": "string"
        }
      },
      "description": ""
    }
  }
}